Server Test - Refactor Remove Duplicates
Remove duplicate code from log tests
